Identity phase 1

+++ Tweezers.Core/Api/Controllers/TweezersObjectController.cs
@@ -12,6 +12,9 @@ namespace Tweezers.Api.Controllers
         [HttpGet("tweezers-schema")]
         public ActionResult<TweezersMultipleResults<TweezersObject>> List([FromQuery] bool internalObj)
         {
+            if (!IsSessionValid())
+                return TweezersUnauthorized();
+
             IEnumerable<TweezersObject> allMetadata = TweezersSchemaFactory.GetAll(includeInternal: internalObj);
             return TweezersOk(TweezersMultipleResults<TweezersObject>.Create(allMetadata));
         }
@@ -19,14 +22,19 @@ namespace Tweezers.Api.Controllers
         [HttpGet("tweezers-schema/{collectionName}")]
         public ActionResult<TweezersObject> Get(string collectionName, [FromQuery] bool internalObj)
         {
-            TweezersObject objectMetadata = TweezersSchemaFactory.Find(collectionName, withInternalObjects: internalObj);
+            if (!IsSessionValid())
+                return TweezersUnauthorized();
 
+            TweezersObject objectMetadata = TweezersSchemaFactory.Find(collectionName, withInternalObjects: internalObj);
             return TweezersOk(objectMetadata);
         }
 
         [HttpPost("tweezers-schema")]
         public ActionResult<TweezersObject> Post([FromBody] TweezersObject data)
         {
+            if (!IsSessionValid())
+                return TweezersUnauthorized();
+
             TweezersObject obj = ReplaceTweezersObject(data);
             return TweezersCreated(obj);
         }
@@ -42,6 +50,9 @@ namespace Tweezers.Api.Controllers
         [HttpPut("tweezers-schema/{collectionName}")]
         public ActionResult<TweezersObject> Patch(string collectionName, [FromBody] TweezersObject data)
         {
+            if (!IsSessionValid())
+                return TweezersUnauthorized();
+
             data.CollectionName = collectionName;
             TweezersObject obj = ReplaceTweezersObject(data);
             return TweezersOk(obj);
@@ -50,6 +61,9 @@ namespace Tweezers.Api.Controllers
         [HttpDelete("tweezers-schema/{collectionName}")]
         public ActionResult<bool> Delete(string collectionName)
         {
+            if (!IsSessionValid())
+                return TweezersUnauthorized();
+
             bool deleted = TweezersSchemaFactory.DeleteObject(collectionName);
             return TweezersOk(deleted);
         }
